[cloud_firestore][wip] migrate to present

+++ cloud_firestore/cloud_firestore_vm/lib/src/firebase/firestore/local/sqlite/sqlite_schema.dart
@@ -13,25 +13,25 @@ part of sqlite_persistence;
 /// changes can be made to the schema by adding a new migration method, bumping the [version], and adding a call to the
 /// migration method from [runMigrations].
 class SQLiteSchema {
-  const SQLiteSchema(this.db);
+  const SQLiteSchema(this.db, this.serializer);
 
   /// The version of the schema. Increase this by one for each migration added to [runMigrations] below.
-  static const int version = 8;
+  static const int version = 11;
 
   // Remove this constant and increment version to enable indexing support
   static const int indexingSupportVersion = version + 1;
 
   final Database db;
+  final LocalSerializer serializer;
 
   /// Runs the migration methods defined in this class, starting at the given version.
-  Future<void> runMigrations(
-      [int fromVersion = 0, int toVersion = version]) async {
+  Future<void> runMigrations([int fromVersion = 0, int toVersion = version]) async {
     // New migrations should be added at the end of the series of `if` statements and should follow the pattern. Make
     // sure to increment `VERSION` and to read the comment below about requirements for new migrations.
 
     if (fromVersion < 1 && toVersion >= 1) {
       await _createV1MutationQueue();
-      await _createV1QueryCache();
+      await _createV1TargetCache();
       await _createV1RemoteDocumentCache();
     }
 
@@ -40,8 +40,8 @@ class SQLiteSchema {
     if (fromVersion < 3 && toVersion >= 3) {
       // Brand new clients don't need to drop and recreate--only clients that have potentially corrupt data.
       if (fromVersion != 0) {
-        await _dropV1QueryCache();
-        await _createV1QueryCache();
+        await _dropV1TargetCache();
+        await _createV1TargetCache();
       }
     }
 
@@ -66,6 +66,31 @@ class SQLiteSchema {
       await _createV8CollectionParentsIndex();
     }
 
+    if (fromVersion < 9 && toVersion >= 9) {
+      if (!await _hasReadTime()) {
+        await _addReadTime();
+      } else {
+        // Index-free queries rely on the fact that documents updated after a query's last limbo
+        // free snapshot version are persisted with their read-time. If a customer upgrades to
+        // schema version 9, downgrades and then upgrades again, some queries may have a last limbo
+        // free snapshot version despite the fact that not all updated document have an associated
+        // read time.
+        await _dropLastLimboFreeSnapshotVersion();
+      }
+    }
+
+    if (fromVersion == 9 && toVersion >= 10) {
+      // Firestore v21.10 contained a regression that led us to disable an assert that is required
+      // to ensure data integrity. While the schema did not change between version 9 and 10, we use
+      // the schema bump to version 10 to clear any affected data.
+      await _dropLastLimboFreeSnapshotVersion();
+    }
+
+    if (fromVersion < 11 && toVersion >= 11) {
+      // Schema version 11 changed the format of canonical IDs in the target cache.
+      await _rewriteCanonicalIds();
+    }
+
     // Adding a new migration? READ THIS FIRST!
     //
     // Be aware that the SDK version may be downgraded then re-upgraded. This means that running your new migration must
@@ -77,8 +102,7 @@ class SQLiteSchema {
     //  from later versions, so migrations that update values cannot assume that existing values have been properly
     //  maintained. Calculate them again, if applicable.
 
-    if (fromVersion < indexingSupportVersion &&
-        toVersion >= indexingSupportVersion) {
+    if (fromVersion < indexingSupportVersion && toVersion >= indexingSupportVersion) {
       if (Persistence.indexingSupportEnabled) {
         await _createLocalDocumentsCollectionIndex();
       }
@@ -89,8 +113,7 @@ class SQLiteSchema {
   /// exist. Use this method to create a set of tables at once.
   ///
   /// If some but not all of the tables exist, an exception will be thrown.
-  Future<void> _ifTablesDontExist(
-      List<String> tables, Future<void> Function() fn) async {
+  Future<void> _ifTablesDontExist(List<String> tables, Future<void> Function() fn) async {
     bool tablesFound = false;
     final String allTables = '[${tables.join(', ')}]';
     for (int i = 0; i < tables.length; i++) {
@@ -99,8 +122,7 @@ class SQLiteSchema {
       if (i == 0) {
         tablesFound = tableFound;
       } else if (tableFound != tablesFound) {
-        final StringBuffer msg = StringBuffer(
-            'Expected all of $allTables to either exist or not, but ');
+        final StringBuffer msg = StringBuffer('Expected all of $allTables to either exist or not, but ');
         if (tablesFound) {
           msg.write('${tables[0]} exists and $table does not');
         } else {
@@ -112,18 +134,15 @@ class SQLiteSchema {
     if (!tablesFound) {
       return fn();
     } else {
-      Log.d('SQLiteSchema',
-          'Skipping migration because all of $allTables already exist');
+      Log.d('SQLiteSchema', 'Skipping migration because all of $allTables already exist');
     }
   }
 
   Future<void> _createV1MutationQueue() async {
-    return _ifTablesDontExist(
-        <String>['mutation_queues', 'mutations', 'document_mutations'],
-        () async {
+    return _ifTablesDontExist(<String>['mutation_queues', 'mutations', 'document_mutations'], () async {
       // A table naming all the mutation queues in the system.
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE TABLE mutation_queues (
             uid                        TEXT PRIMARY KEY,
@@ -136,7 +155,7 @@ class SQLiteSchema {
 
       // All the mutation batches in the system, partitioned by user.
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE TABLE mutations (
             uid       TEXT,
@@ -151,7 +170,7 @@ class SQLiteSchema {
       // A manually maintained index of all the mutation batches that affect a given document key. The rows in this
       // table are references based on the contents of mutations.mutations.
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE TABLE document_mutations (
             uid      TEXT,
@@ -167,16 +186,15 @@ class SQLiteSchema {
 
   /// Note: as of this migration, [last_acknowledged_batch_id] is no longer used by the code.
   Future<void> _removeAcknowledgedMutations() async {
-    final List<Map<String, dynamic>> data = await db
-        .query('SELECT uid, last_acknowledged_batch_id FROM mutation_queues');
+    final List<Map<String, dynamic>> data =
+        await db.query('SELECT uid, last_acknowledged_batch_id FROM mutation_queues');
 
     for (Map<String, dynamic> row in data) {
       final String uid = row['uid'];
       final int lastAcknowledgedBatchId = row['last_acknowledged_batch_id'];
 
       final List<Map<String, dynamic>> rows = await db.query(
-          'SELECT batch_id FROM mutations WHERE uid = ? AND batch_id <= ?',
-          <dynamic>[uid, lastAcknowledgedBatchId]);
+          'SELECT batch_id FROM mutations WHERE uid = ? AND batch_id <= ?', <dynamic>[uid, lastAcknowledgedBatchId]);
 
       for (Map<String, dynamic> row in rows) {
         await _removeMutationBatch(uid, row['batch_id']);
@@ -185,23 +203,19 @@ class SQLiteSchema {
   }
 
   Future<void> _removeMutationBatch(String uid, int batchId) async {
-    final int deleted = await db.delete(
-        'DELETE FROM mutations WHERE uid = ? AND batch_id = ?',
-        <dynamic>[uid, batchId]);
+    final int deleted =
+        await db.delete('DELETE FROM mutations WHERE uid = ? AND batch_id = ?', <dynamic>[uid, batchId]);
     hardAssert(deleted != 0, 'Mutation batch ($uid, $batchId) did not exist');
 
     // Delete all index entries for this batch
-    return db.delete(
-        'DELETE FROM document_mutations WHERE uid = ? AND batch_id = ?',
-        <dynamic>[uid, batchId]);
+    return db.delete('DELETE FROM document_mutations WHERE uid = ? AND batch_id = ?', <dynamic>[uid, batchId]);
   }
 
-  Future<void> _createV1QueryCache() async {
-    return _ifTablesDontExist(
-        <String>['targets', 'target_globals', 'target_documents'], () async {
+  Future<void> _createV1TargetCache() async {
+    return _ifTablesDontExist(<String>['targets', 'target_globals', 'target_documents'], () async {
       // A cache of targets and associated metadata
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE TABLE targets (
             target_id                   INTEGER PRIMARY KEY,
@@ -217,7 +231,7 @@ class SQLiteSchema {
           );
 
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE INDEX query_targets
             ON targets (canonical_id, target_id);
@@ -227,7 +241,7 @@ class SQLiteSchema {
 
       // Global state tracked across all queries, tracked separately
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE TABLE target_globals (
             highest_target_id                    INTEGER,
@@ -241,7 +255,7 @@ class SQLiteSchema {
 
       // A Mapping table between targets, document paths
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE TABLE target_documents (
             target_id       INTEGER,
@@ -254,7 +268,7 @@ class SQLiteSchema {
 
       // The document_targets reverse mapping table is just an index on target_documents.
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE INDEX document_targets
             ON target_documents (path, target_id);
@@ -264,7 +278,7 @@ class SQLiteSchema {
     });
   }
 
-  Future<void> _dropV1QueryCache() async {
+  Future<void> _dropV1TargetCache() async {
     // This might be overkill, but if any future migration drops these, it's possible we could try dropping tables that
     // don't exist.
     if (await _tableExists('targets')) {
@@ -282,7 +296,7 @@ class SQLiteSchema {
     return _ifTablesDontExist(<String>['remote_documents'], () async {
       // A cache of documents obtained from the server.
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE TABLE remote_documents (
             path     TEXT PRIMARY KEY,
@@ -302,7 +316,7 @@ class SQLiteSchema {
     return _ifTablesDontExist(<String>['collection_index'], () async {
       // A per-user, per-collection index for cached documents indexed by a single field's name and value.
       await db.query(
-          // @formatter:off
+        // @formatter:off
           '''
           CREATE TABLE collection_index (
             uid              TEXT,
@@ -325,7 +339,7 @@ class SQLiteSchema {
 
     if (!targetGlobalExists) {
       await db.execute(
-          // @formatter:off
+        // @formatter:off
           '''
           INSERT INTO target_globals
           (highest_target_id, highest_listen_sequence_number, last_remote_snapshot_version_seconds,
@@ -339,13 +353,12 @@ class SQLiteSchema {
 
   Future<void> _addTargetCount() async {
     if (!(await _tableContainsColumn('target_globals', 'target_count'))) {
-      await db.execute(
-          'ALTER TABLE target_globals ADD COLUMN target_count INTEGER');
+      await db.execute('ALTER TABLE target_globals ADD COLUMN target_count INTEGER');
     }
     // Even if the column already existed, rerun the data migration to make sure it's correct.
     final int count = await _rowNumber('targets');
     return db.execute(
-        // @formatter:off
+      // @formatter:off
         '''
           UPDATE target_globals
           SET target_count=?
@@ -356,8 +369,41 @@ class SQLiteSchema {
 
   Future<void> _addSequenceNumber() async {
     if (!(await _tableContainsColumn('target_documents', 'sequence_number'))) {
-      return db.execute(
-          'ALTER TABLE target_documents ADD COLUMN sequence_number INTEGER');
+      return db.execute('ALTER TABLE target_documents ADD COLUMN sequence_number INTEGER');
+    }
+  }
+
+  Future<bool> _hasReadTime() async {
+    final bool hasReadTimeSeconds = await _tableContainsColumn('remote_documents', 'read_time_seconds');
+    final bool hasReadTimeNanos = await _tableContainsColumn('remote_documents', 'read_time_nanos');
+
+    hardAssert(
+      hasReadTimeSeconds == hasReadTimeNanos,
+      'Table contained just one of read_time_seconds or read_time_nanos',
+    );
+
+    return hasReadTimeSeconds && hasReadTimeNanos;
+  }
+
+  Future<void> _addReadTime() async {
+    await db.execute('ALTER TABLE remote_documents ADD COLUMN read_time_seconds INTEGER');
+    await db.execute('ALTER TABLE remote_documents ADD COLUMN read_time_nanos INTEGER');
+  }
+
+  Future<void> _dropLastLimboFreeSnapshotVersion() async {
+    final List<Map<String, dynamic>> rows = await db.query('SELECT target_id, target_proto FROM targets');
+
+    for (Map<String, dynamic> row in rows) {
+      final int targetId = row['target_id'];
+      final Uint8List targetProtoBytes = Uint8List.fromList(row['target_proto']);
+
+      final proto.Target targetProto = proto.Target.fromBuffer(targetProtoBytes) //
+        ..clearLastLimboFreeSnapshotVersion();
+
+      await db.execute(
+        'UPDATE targets SET target_proto = ? WHERE target_id = ?',
+        <dynamic>[targetProto.writeToBuffer(), targetId],
+      );
     }
   }
 
@@ -367,7 +413,7 @@ class SQLiteSchema {
   Future<void> _ensureSequenceNumbers() async {
     // Get the current highest sequence number
     final List<Map<String, dynamic>> sequenceNumberQuery = await db.query(
-        // @formatter:off
+      // @formatter:off
         '''
           SELECT highest_listen_sequence_number
           FROM target_globals
@@ -375,12 +421,11 @@ class SQLiteSchema {
         '''
         // @formatter:on
         );
-    final int sequenceNumber =
-        sequenceNumberQuery.first['highest_listen_sequence_number'];
+    final int sequenceNumber = sequenceNumberQuery.first['highest_listen_sequence_number'];
     assert(sequenceNumber != null, 'Missing highest sequence number');
 
     final List<Map<String, dynamic>> untaggedDocumentsQuery = await db.query(
-        // @formatter:off
+      // @formatter:off
         '''
           SELECT RD.path
           FROM remote_documents AS RD
@@ -419,7 +464,7 @@ class SQLiteSchema {
         // parent path will be an empty path in the case of root-level
         // collections.
         await db.execute(
-            // @formatter:off
+          // @formatter:off
             '''
               CREATE TABLE collection_parents(
                   collection_id TEXT,
@@ -455,7 +500,7 @@ class SQLiteSchema {
 
     // Index existing remote documents.
     final List<Map<String, dynamic>> remoteDocumentsQuery = await db.query(
-        // @formatter:off
+      // @formatter:off
         '''
           SELECT path
           FROM remote_documents;
@@ -470,7 +515,7 @@ class SQLiteSchema {
 
     // Index existing mutations.
     final List<Map<String, dynamic>> documentMutationsQuery = await db.query(
-        // @formatter:off
+      // @formatter:off
         '''
           SELECT path
           FROM document_mutations;
@@ -484,6 +529,20 @@ class SQLiteSchema {
     }
   }
 
+  Future<void> _rewriteCanonicalIds() async {
+    final List<Map<String, dynamic>> rows = await db.query('SELECT target_id, target_proto FROM targets');
+    for (final Map<String, dynamic> row in rows) {
+      final int targetId = row['target_id'];
+      final Uint8List targetProtoBytes = Uint8List.fromList(row['target_proto']);
+
+      final proto.Target targetProto = proto.Target.fromBuffer(targetProtoBytes);
+      final TargetData targetData = serializer.decodeTargetData(targetProto);
+      final String updatedCanonicalId = targetData.target.canonicalId;
+      await db
+          .execute('UPDATE targets SET canonical_id  = ? WHERE target_id = ?', <dynamic>[updatedCanonicalId, targetId]);
+    }
+  }
+
   Future<bool> _tableContainsColumn(String table, String column) async {
     final List<String> columns = await getTableColumns(table);
     return columns.contains(column);
@@ -491,19 +550,17 @@ class SQLiteSchema {
 
   @visibleForTesting
   Future<List<String>> getTableColumns(String table) async {
-    final List<Map<String, dynamic>> data =
-        await db.query('PRAGMA table_info($table);');
+    final List<Map<String, dynamic>> data = await db.query('PRAGMA table_info($table);');
     return data.map<String>((Map<String, dynamic> row) => row['name']).toList();
   }
 
   Future<bool> _tableExists(String table) async {
-    final List<Map<String, dynamic>> data = await db.query(
-        'SELECT 1=1 FROM sqlite_master WHERE tbl_name = ?', <String>[table]);
+    final List<Map<String, dynamic>> data =
+        await db.query('SELECT 1=1 FROM sqlite_master WHERE tbl_name = ?', <String>[table]);
     return data.isNotEmpty;
   }
 
   Future<int> _rowNumber(String tableName) async {
-    return (await db.query('SELECT Count(*) as count FROM $tableName;'))
-        .first['count'];
+    return (await db.query('SELECT Count(*) as count FROM $tableName;')).first['count'];
   }
 }
